Extracts base class out of OAuthBackend
